# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Simple REST API developed developed for submission assigment 2 Bootcamp at Hacktive8",
        "title": "Hacktiv8 Assigment 2",
        "contact": {
            "name": "Teguh Ainul Darajat",
            "email": "teguh@email.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/orders": {
            "get": {
                "description": "API for Get All order data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/params.OrderResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "API for create order",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "parameters": [
                    {
                        "description": "Create order",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/params.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/params.OrderResponse"
                        }
                    }
                }
            }
        },
        "/orders/:id": {
            "put": {
                "description": "API for update order data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "parameters": [
                    {
                        "description": "Update Order",
                        "name": "Body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/params.Order"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/params.OrderResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "API for Delete order data",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/params.OrderResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "params.Item": {
            "type": "object",
            "properties": {
                "OrderID": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "item_code": {
                    "type": "string"
                },
                "quantity": {
                    "type": "integer"
                }
            }
        },
        "params.Order": {
            "type": "object",
            "properties": {
                "customer_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/params.Item"
                    }
                },
                "ordered_at": {
                    "type": "string"
                }
            }
        },
        "params.OrderResponse": {
            "type": "object",
            "properties": {
                "additional_info": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "payload": {
                    "$ref": "#/definitions/params.Order"
                },
                "status": {
                    "type": "integer"
                }
            }
        }
    }
}